Added new hidden option --devel to global(1). This options is to test the built-in sort filter 'gtags --sort' for a while, and will be removed after the test.
When the --devel option is specified, global(1) use 'gtags --sort [--unique]'
as a sort filter instead of 'gnusort -k 1,1 -k 3,3 -k 2,2n [-u]' in the
following case:

        global -x xxx
        global -xr xxx
        global -xs xxx

It seems that the built-in sort command works well, and it is faster than
gnusort in case of almost. But I cannot determine whether we should switch
to it yet.